move pseudo under util
