--- logprep/pseudo/commands/version.py
@@ -1,11 +1,0 @@
-import click
-
-from .. import __version__
-
-
-@click.command()
-def print_version():
-    """
-    Print version string
-    """
-    print(__version__)
