+++ logprep/util/pseudo/commands/generate_keys.py
@@ -1,11 +1,13 @@
 import click
-from ..keygenerator import generate_rsa_key
+
+from logprep.util.pseudo.keygenerator import generate_rsa_key
 
 
 @click.command()
 @click.argument("key-length", default="1024", type=int)
 @click.option("-f", "--file")
 def generate(key_length: int, file: str):
+    """Generate RSA keys for pseudonymization."""
     priv_key, pub_key = generate_rsa_key.generate_keys(key_length=key_length)
     if not file:
         print(priv_key.decode("utf8"))
